feat: make stack 1byte per address

+++ src/stack.h
@@ -5,7 +5,7 @@
 
 typedef struct
 {
-    uint16_t *data;
+    uint8_t *data;
     uint16_t size;
     uint16_t sp;
 } Stack;
@@ -14,9 +14,9 @@ Stack *stack_create(uint16_t size);
 
 void stack_free(Stack *stack);
 
-uint16_t stack_pop(Stack *stack);
+uint8_t stack_pop(Stack *stack);
 
-void stack_push(Stack *stack, uint16_t value);
+void stack_push(Stack *stack, uint8_t value);
 
 void stack_print(Stack *stack);
